+++ src/main/java/ru/evotor/framework/core/BroadcastEventReceiver.kt
@@ -1,11 +1,11 @@
-package ru.evotor.framework.core.action.broadcast
+package ru.evotor.framework.core
 
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 
-abstract class AbstractBroadcastReceiver : BroadcastReceiver() {
+abstract class BroadcastEventReceiver : BroadcastReceiver() {
 
     final override fun onReceive(context: Context, intent: Intent) {
         if (intent.action != null && intent.extras != null) {
