+++ src/main/java/ru/evotor/framework/device/cash_drawer/event/handler/receiver/CashDrawerBroadcastReceiver.kt
@@ -1,14 +1,16 @@
-package ru.evotor.framework.core.action.broadcast
+package ru.evotor.framework.device.cash_drawer.event.handler.receiver
 
 import android.content.Context
 import android.os.Bundle
+import ru.evotor.framework.core.RequiresIntentAction
+import ru.evotor.framework.core.BroadcastEventReceiver
 import ru.evotor.framework.core.action.event.cash_drawer.CashDrawerOpenEvent
 
 /**
  * Широковещательный приёмник событий денежного ящика.
  * @see <a href="https://developer.evotor.ru/docs/beta/doc_java_broadcastreceiver.html">Использование широковещательного приёмника</a>
  */
-abstract class CashDrawerBroadcastReceiver : AbstractBroadcastReceiver() {
+abstract class CashDrawerBroadcastReceiver : BroadcastEventReceiver() {
 
     /**
      * Обработчик событий открытия денежного ящика.
