+++ src/main/java/ru/evotor/framework/device/scanner/event/handler/receiver/ScannerBroadcastReceiver.kt
@@ -1,16 +1,18 @@
-package ru.evotor.framework.core.action.broadcast
+package ru.evotor.framework.device.scanner.event.handler.receiver
 
 import android.annotation.SuppressLint
 import android.content.Context
 import android.os.Bundle
 import android.support.annotation.RequiresPermission
-import ru.evotor.framework.core.action.event.scanner.BarcodeReceivedEvent
+import ru.evotor.framework.core.RequiresIntentAction
+import ru.evotor.framework.core.BroadcastEventReceiver
+import ru.evotor.framework.device.scanner.event.BarcodeReceivedEvent
 
 /**
  * Широковещательный приёмник событий сканера штрихкодов.
  * @see <a href="https://developer.evotor.ru/docs/beta/doc_java_barcode_scanner.html">Использование широковещательного приёмника</a>
  */
-abstract class ScannerBroadcastReceiver : AbstractBroadcastReceiver() {
+abstract class ScannerBroadcastReceiver : BroadcastEventReceiver() {
 
     /**
      * Обработчик событий получения штрихкода.
