+++ src/main/java/ru/evotor/framework/receipt/event/handler/receiver/BuybackReceiptBroadcastReceiver.kt
@@ -1,6 +1,7 @@
-package ru.evotor.framework.core.action.broadcast
+package ru.evotor.framework.receipt.event.handler.receiver
 
 import android.content.Context
+import ru.evotor.framework.core.RequiresIntentAction
 import ru.evotor.framework.core.action.event.receipt.position_edited.PositionAddedEvent
 import ru.evotor.framework.core.action.event.receipt.position_edited.PositionEditedEvent
 import ru.evotor.framework.core.action.event.receipt.position_edited.PositionRemovedEvent
@@ -9,10 +10,10 @@ import ru.evotor.framework.core.action.event.receipt.receipt_edited.ReceiptClose
 import ru.evotor.framework.core.action.event.receipt.receipt_edited.ReceiptOpenedEvent
 
 /**
- * Широковещательный приёмник событий чека возврата.
+ * Широковещательный приёмник событий чека возврата покупки.
  * @see <a href="https://developer.evotor.ru/docs/beta/doc_java_broadcastreceiver.html">Использование широковещательного приёмника</a>
  */
-open class PaybackReceiptBroadcastReceiver : ReceiptBroadcastReceiver(
+open class BuybackReceiptBroadcastReceiver : ReceiptBroadcastReceiver(
         ACTION_RECEIPT_OPENED,
         ACTION_POSITION_ADDED,
         ACTION_POSITION_EDITED,
@@ -59,17 +60,17 @@ open class PaybackReceiptBroadcastReceiver : ReceiptBroadcastReceiver(
 
     companion object {
 
-        const val ACTION_RECEIPT_OPENED = "evotor.intent.action.receipt.payback.OPENED"
+        const val ACTION_RECEIPT_OPENED = "evotor.intent.action.receipt.buyback.OPENED"
 
-        const val ACTION_POSITION_ADDED = "evotor.intent.action.receipt.payback.POSITION_ADDED"
+        const val ACTION_POSITION_ADDED = "evotor.intent.action.receipt.buyback.POSITION_ADDED"
 
-        const val ACTION_POSITION_EDITED = "evotor.intent.action.receipt.payback.POSITION_EDITED"
+        const val ACTION_POSITION_EDITED = "evotor.intent.action.receipt.buyback.POSITION_EDITED"
 
-        const val ACTION_POSITION_REMOVED = "evotor.intent.action.receipt.payback.POSITION_REMOVED"
+        const val ACTION_POSITION_REMOVED = "evotor.intent.action.receipt.buyback.POSITION_REMOVED"
 
-        const val ACTION_RECEIPT_CLEARED = "evotor.intent.action.receipt.payback.CLEARED"
+        const val ACTION_RECEIPT_CLEARED = "evotor.intent.action.receipt.buyback.CLEARED"
 
-        const val ACTION_RECEIPT_CLOSED = "evotor.intent.action.receipt.payback.CLOSED"
+        const val ACTION_RECEIPT_CLOSED = "evotor.intent.action.receipt.buyback.CLOSED"
 
     }
