+++ src/main/java/ru/evotor/framework/receipt/event/handler/receiver/SellReceiptBroadcastReceiver.kt
@@ -1,6 +1,7 @@
-package ru.evotor.framework.core.action.broadcast
+package ru.evotor.framework.receipt.event.handler.receiver
 
 import android.content.Context
+import ru.evotor.framework.core.RequiresIntentAction
 import ru.evotor.framework.core.action.event.receipt.position_edited.PositionAddedEvent
 import ru.evotor.framework.core.action.event.receipt.position_edited.PositionEditedEvent
 import ru.evotor.framework.core.action.event.receipt.position_edited.PositionRemovedEvent
